"Changed type annotation of function parameter 'n' from 'int' to 'float'."

+++ 0x00-python_variable_annotations/2-floor.py
@@ -2,6 +2,6 @@
 """Define a type-annotated function floor"""
 
 
-def floor(n: int) -> int:
+def floor(n: float) -> int:
     """Return the floor of a float number"""
     return int(n // 1)
